requirements.txt
pygame-ce
tensorflow
matplotlib